+++ server/libs/automation/automation-knowledge-base/automation-knowledge-base-service/src/main/java/com/bytechef/automation/knowledgebase/service/KnowledgeBaseVectorStoreMetadataService.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.bytechef.automation.knowledgebase.config;
+package com.bytechef.automation.knowledgebase.service;
 
 import static com.bytechef.automation.knowledgebase.constant.KnowledgeBaseConstants.METADATA_TAG_NAMES;
 
@@ -34,9 +34,9 @@ import tools.jackson.databind.ObjectMapper;
  *
  * @author Marko Kriskovic
  */
-public final class KnowledgeBaseVectorStoreMetadataUpdater {
+public final class KnowledgeBaseVectorStoreMetadataService {
 
-    private static final Logger logger = LoggerFactory.getLogger(KnowledgeBaseVectorStoreMetadataUpdater.class);
+    private static final Logger logger = LoggerFactory.getLogger(KnowledgeBaseVectorStoreMetadataService.class);
     private static final Pattern SAFE_TABLE_NAME_PATTERN = Pattern.compile("^[a-zA-Z_][a-zA-Z0-9_.]*$");
 
     private final JdbcTemplate pgVectorJdbcTemplate;
@@ -44,7 +44,7 @@ public final class KnowledgeBaseVectorStoreMetadataUpdater {
     private final String fullTableName;
 
     @SuppressFBWarnings("EI2")
-    public KnowledgeBaseVectorStoreMetadataUpdater(
+    public KnowledgeBaseVectorStoreMetadataService(
         JdbcTemplate pgVectorJdbcTemplate, ObjectMapper objectMapper, String fullTableName) {
 
         if (!SAFE_TABLE_NAME_PATTERN.matcher(fullTableName)
